pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "quant-stack-video"
version = "0.1.0"
description = "Infinite layered video generation via quantization stacking, VACE temporal extension, and SVI error recycling"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
keywords = ["video-generation", "quantization", "diffusion", "wan", "vace", "bitsandbytes"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "diffusers>=0.30.0",
    "transformers>=4.40.0",
    "accelerate>=0.30.0",
    "bitsandbytes>=0.43.0",
    "safetensors>=0.4.0",
    "huggingface-hub>=0.23.0",
    "numpy>=1.26.0",
    "Pillow>=10.0.0",
    "imageio>=2.34.0",
    "imageio-ffmpeg>=0.4.9",
    "scikit-image>=0.22.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "tqdm>=4.66.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
perceptual = ["lpips>=0.1.4"]
dev = ["pytest>=8.0.0", "ruff>=0.4.0"]

[project.scripts]
quant-generate = "scripts.generate:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
